+++ webfiori/framework/cli/commands/AddCommand.php
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace webfiori\framework\cli\writers;
+namespace webfiori\framework\cli\commands;
 
 use Exception;
 use webfiori\database\ConnectionInfo;
@@ -30,6 +30,7 @@ use webfiori\framework\ConfigController;
 use webfiori\framework\DB;
 use webfiori\framework\mail\SMTPAccount;
 use webfiori\framework\WebFioriApp;
+use webfiori\framework\cli\CLICommand;
 
 /**
  * A command which is used to add a database connection or SMTP account.
